requirements_agent.txt
requests
python-dotenv
pyyaml
websockets
simple-websocket
